+++ buildpacks/nodejs-yarn/CHANGELOG.md
@@ -2,10 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- Added Yarn version 4.0.0-rc.53.
+- Added Yarn version 4.0.0-rc.52.
+- Added Yarn version 3.6.4.
+## [1.1.6] - 2023-09-25
+
+- No changes.
+
+## [1.1.5] - 2023-09-19
+
+- Added Yarn version 4.0.0-rc.51.
+- Added Yarn version 4.0.0-rc.50.
+- Added Yarn version 4.0.0-rc.49.
+- Added Yarn version 3.6.3.
+- Added Yarn version 3.6.2.
 
 ## [1.1.4] - 2023-08-10
 
@@ -128,7 +143,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Changelog entry for first release ([#1](https://github.com/heroku/nodejs-yarn-buildpack/pull/1))
 
-[unreleased]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.4...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.6...HEAD
+[1.1.6]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.5...v1.1.6
+[1.1.5]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.3...v1.1.4
 [1.1.3]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.1...v1.1.2