+++ meta-buildpacks/nodejs-function/CHANGELOG.md
@@ -2,10 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.6] - 2023-09-25
+
+- Updated `heroku/nodejs-engine` to `1.1.6`.
+- Updated `heroku/nodejs-function-invoker` to `1.1.6`.
+- Updated `heroku/nodejs-npm` to `1.1.6`.
+
+## [1.1.5] - 2023-09-19
+
+- Updated `heroku/nodejs-engine` to `1.1.5`.
+- Updated `heroku/nodejs-function-invoker` to `1.1.5`.
+- Updated `heroku/nodejs-npm` to `1.1.5`.
 
 ## [1.1.4] - 2023-08-10
 
@@ -269,7 +281,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * Switch from the Riff based invoker buildpacks to `heroku/nodejs-function` ([#48](https://github.com/heroku/buildpacks-node/pull/48))
 
-[unreleased]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.4...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.6...HEAD
+[1.1.6]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.5...v1.1.6
+[1.1.5]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.3...v1.1.4
 [1.1.3]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/heroku/buildpacks-nodejs/compare/v1.1.1...v1.1.2